+++ packages/engine/src/networking/schema/worldStateSchema.ts
@@ -1,4 +1,4 @@
-import { float32, Model, Schema, string, uint32, uint64, uint8 } from "superbuffer";
+import { float32, Model, Schema, string, uint32, uint8 } from "superbuffer";
 import { PacketWorldState, WorldStateInterface } from "../interfaces/WorldState";
 //import { inputKeyArraySchema } from "./clientInputSchema";
 
@@ -46,7 +46,7 @@ const clientDisconnectedSchema = new Schema({
 
 const transformSchema = new Schema({
     networkId: uint32,
-    snapShotTime: uint64,
+    snapShotTime: uint32,
     x: float32,
     y: float32,
     z: float32,
@@ -59,7 +59,7 @@ const transformSchema = new Schema({
 const snapshotSchema = new Schema({
   id: string,
   state: [transformSchema],
-  time: uint64
+  time: uint32
 })
 
 const createNetworkObjectSchema = new Schema({
@@ -85,7 +85,7 @@ const worldStateSchema = new Schema({
     createObjects: [createNetworkObjectSchema],
     destroyObjects: [destroyNetworkObjectSchema],
     inputs: [inputKeyArraySchema],
-    tick: uint64,
+    tick: uint32,
     transforms: [transformSchema]
 });
